extract super class
